// CommunicationSclEditor.d.ts
import { LitElement, TemplateResult } from 'lit';
import { MdFab } from '@scopedelement/material-web/fab/MdFab.js';
import { MdIcon } from '@scopedelement/material-web/icon/MdIcon.js';
import { MdIconButton } from '@scopedelement/material-web/iconbutton/MdIconButton.js';
import { MdOutlinedTextField } from '@scopedelement/material-web/textfield/MdOutlinedTextField.js';
import { Point } from './foundation/sldUtil.js';
import { IED, Connection } from './foundation/types.js';
declare const CommunicationSclEditor_base: typeof LitElement & import("@open-wc/scoped-elements/lit-element.js").ScopedElementsHostConstructor;
export declare class CommunicationSclEditor extends CommunicationSclEditor_base {
    static scopedElements: {
        'md-fab': typeof MdFab;
        'md-icon': typeof MdIcon;
        'md-outlined-textfield': typeof MdOutlinedTextField;
        'md-icon-button': typeof MdIconButton;
    };
    substation: Element;
    gridSize: number;
    connections: Connection[];
    get ieds(): IED[];
    filterReport: boolean;
    filterGOOSE: boolean;
    filterSMV: boolean;
    selectedIed?: Element;
    filterRcv: boolean;
    filterSend: boolean;
    sourceIEDFilter: string;
    targetIEDFilter: string;
    cbNameFilter: string;
    showFilterBox: boolean;
    editMode: boolean;
    showLabel: boolean;
    placing?: Element;
    placingLabel?: Element;
    placingOffset: Point;
    mouseX: number;
    mouseY: number;
    mouseX2: number;
    mouseY2: number;
    linkedEquipments: Element[];
    get idle(): boolean;
    sld: SVGGraphicsElement;
    container: HTMLDivElement;
    svgCoordinates(clientX: number, clientY: number): Point;
    reset(): void;
    placeLabel(element: Element, x: number, y: number): void;
    startPlacingLabel(element: Element | undefined, offset?: Point): void;
    placeElement(element: Element, x: number, y: number): void;
    startPlacing(element: Element | undefined, offset?: Point): void;
    onWheelZoom(evt: WheelEvent): void;
    clearFilter(): void;
    activeFilter(): boolean;
    filterCbName(conn: Connection): boolean;
    filterTargetIED(conn: Connection): boolean;
    filterSourceIED(conn: Connection): boolean;
    filterConnections(conn: Connection): boolean;
    resetIedSelection(): void;
    selectIED(ied: IED): void;
    constructor();
    renderedLabelPosition(element: Element): Point;
    renderLabel(element: Element): TemplateResult<2>;
    renderedPosition(element: Element): Point;
    renderIED(ied: IED): TemplateResult<2>;
    renderFilterBox(): TemplateResult;
    renderFilterFab(): TemplateResult;
    renderService(controlBlock: string): TemplateResult[];
    renderInfoBox(): TemplateResult;
    render(): TemplateResult<1>;
    static styles: import("lit").CSSResult;
}
export {};